Add IframeExecutionService tests
